scr/requirements.txt
fastapi
uvicorn
asyncio
ultralytics
numpy
opencv-contrib-python
websockets
fastapi[standard]
uvicorn[standard]
python-multipart
pillow
aiofiles
tqdm
matplotlib
seaborn
pandas
requests
typing
